refactor: Consolidate all handlers into handlers.go for upstream alignment
This commit reverts previous modularization efforts (splitting autoreply
and Google Contacts features into separate files) and consolidates all
handler logic, type definitions, and relevant helpers back into
`handlers.go`. This is done to align the codebase structure more
closely with the upstream repository (asternic/wuzapi) and minimize
future merge conflicts, as per your feedback.

Specific changes:
- All struct definitions from `autoreply_types.go` and
  `google_contacts_types.go` were moved back to `handlers.go`.
  The `_types.go` files were removed.
- All handler functions and their private helpers from
  `autoreply_handlers.go` and `google_contacts_integration.go`
  were moved back to `handlers.go`. The separate handler files
  were removed.
- Route definitions from `autoreply_routes.go` were moved back into
  the main `routes()` function in `routes.go`. The
  `autoreply_routes.go` file was removed, and the call to its
  registration function was removed from `routes.go`.

Helper function consolidation:
- `Values` struct and `updateUserInfo` function (with signature
  accepting `interface{}` for context value) are now in `handlers.go`,
  aligning with upstream. Removed from `helpers.go`.
- `Find` function (returning `int, bool`) remains in `helpers.go` as
  it's not present in upstream `handlers.go`.
- `parseJID` function remains in `wmiau.go`.
- `s.Respond` is used as the primary response method in `handlers.go`.

Error corrections:
- Corrected usage of `Find` in `handlers.go` (in `Connect`,
  `UpdateWebhook`, `SetWebhook`, `AddUser` handlers) to properly
  handle its two return values.
- Corrected calls to `updateUserInfo` in `handlers.go` to pass
  the context value directly, as the function now handles type assertion.
- Created `constants.go` for `supportedEventTypes` (formerly `messageTypes`)
  as per upstream structure, and updated references.

Imports were reviewed and cleaned in all affected files.
This set of changes aims to make the codebase structure, particularly
`handlers.go`, more similar to the upstream repository.
