+++ constants.go
@@ -1,0 +1,4 @@
+package main
+
+// supportedEventTypes lists the valid event types for webhooks.
+var supportedEventTypes = []string{"Message", "ReadReceipt", "Presence", "HistorySync", "ChatPresence", "All"}
